+ User can select a following 'notification description'(Suggested by "Rob D Johansen"<Rob_D_Johansen@notes.seagate.com>). [Last, First] *E1 Year     : [Jeong, JaeMok] B 1970 (31) [First Last] *E1 Year      : [JaeMok Jeong] B 1970 (31) Event - Last, First Year   : Birthday - Jeong, JaeMok 1970 (31) Event - First Last Year    : Birthday - JaeMok Jeong 1970 (31) * First Last Year          : * JaeMok Jeong 1970 (31)
+ ToDo notification support (todo category, priority)
+ Graffiti searching(Suggested by "Vince Rice"
<vrice@solidrocksystems.com>).
If you write character or number in graffiti area, the
matching row is highlighted. It can be useful with 'sort'.
For example, if you write 'a' in graffiti, the first record
beginining with 'a' or 'A' is highlighted. If you write '1'
in graffiti, the nearest incoming record after January is
highlighted.
+ Progress dialog is displayed in doing a long notification
work(Suggested by "Hugh Harries" <harries@cicy.mx>).
+ In preference menu, 'sort by' control is disappeared. You can
sort the happydays entries only by graffiti 'abc' and '123'.
+ Preference record has the version number. You don't need to
uninstall the old version any more.
+ Code optimize and a minor bug fix in displaying datebk3 icon.
+ Fix some minor bug(eg. boundary check error in recurring
event).